@@ -1,8 +1,8 @@
-require_relative('CommandTemplate')
+require_relative 'CommandTemplate'
 
 class NullCommand < CommandTemplate
-    def execute token
-      token.add_header({:header => 'NullCommand', :value => 'OK'})
-      token
-    end
+  def execute(token)
+    token.add_header(header: 'NullCommand', value: 'OK')
+    token
+  end
 end
