@@ -1,13 +1,23 @@
 class CommandFactory
-  def build command
-    if command['command'] == 'erb'
-      return ErbCommand.new command
-    elsif command['command'] == 'email'
-      return EmailCommand.new command
-    elsif command['command'] == 'concatenate'
-      return ConcatenateCommand.new command
+  COMMAND_MAP = {
+    'erb' => ErbCommand,
+    'email' => EmailCommand,
+    'concatenate' => ConcatenateCommand
+  }.freeze
+
+  def build(command)
+    command_type = command['command']
+    command_class = COMMAND_MAP[command_type]
+
+    if command_class
+      command_class.new(command)
     else
-      return NullCommand.new command
+      LOGGER.warn("Unknown command type '#{command_type}', using NullCommand")
+      NullCommand.new(command)
     end
+  end
+
+  def self.available_commands
+    COMMAND_MAP.keys
   end
 end
