@@ -1,23 +1,42 @@
-require_relative('CommandTemplate')
+require_relative 'CommandTemplate'
 require 'mail'
 
-Mail.defaults do
-  #delivery_method :smtp, address: "localhost", port: 25
-  delivery_method :test
-end
-
 class EmailCommand < CommandTemplate
-    def execute token
-      mail = Mail.new
-      mail[:from] = 'email@example.com'
-      mail[:to] = @command['data']['to']
-      mail[:subject] = @command['data']['subject']
-      mail[:body] = token.get_body
+  def initialize(command)
+    super
+    validate_required_data_keys!('to', 'subject')
+    validate_email_address!(command['data']['to'])
+  end
 
-      mail.deliver
-
-      token.add_header({:header => 'EmailCommand', :value => 'OK'})
-      token.set_body mail.to_s
-      token
+  def execute(token)
+    mail = Mail.new do
+      from    email_config[:from]
+      to      command['data']['to']
+      subject command['data']['subject']
+      body    token.get_body
     end
+
+    mail.deliver!
+
+    token.add_header(header: 'EmailCommand', value: 'OK')
+    token.set_body(mail.to_s)
+    token
+  rescue => e
+    LOGGER.error("Failed to send email: #{e.message}")
+    raise StandardError, "Email delivery failed: #{e.message}"
+  end
+
+  private
+
+  def email_config
+    @email_config ||= begin
+      require_relative '../config/config'
+      Config.email
+    end
+  end
+
+  def validate_email_address!(email)
+    email_regex = /\A[\w+\-.]+@[a-z\d\-]+(\.[a-z\d\-]+)*\.[a-z]+\z/i
+    raise ArgumentError, "Invalid email address: #{email}" unless email =~ email_regex
+  end
 end
